requirements.txt
# vLLM Server Requirements
# Python 3.8-3.11 지원

# 핵심 ML/DL 라이브러리 (버전 호환성 수정)
torch>=2.1.0
torchvision
transformers>=4.45.0
accelerate>=0.24.0

# vLLM 및 관련 의존성
vllm>=0.6.0
ray[default]>=2.7.0
bitsandbytes>=0.46.1

# 웹 프레임워크
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# 이미지 처리
Pillow>=10.0.0
opencv-python>=4.8.0

# 수치 계산
numpy>=1.24.0
scipy>=1.11.0

# 문서 처리
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# 유틸리티
python-dotenv>=1.0.0
requests>=2.31.0

# GPU 가속 (선택적)
# flash-attn>=2.3.0  # vLLM이 자동으로 처리
